requirements.txt
# Core Dependencies
streamlit
pandas
numpy

# Data Fetching & APIs
requests
yfinance
finnhub-python

# Environment Variables
python-dotenv

# Visualization
plotly

# LangChain & AI/LLM
langchain
langchain-openai
langchain-google-genai
openai
google-generativeai

# Optional but Recommended
python-dateutil
aiohttp